requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
iniconfig==2.0.0
newsapi-python==0.2.7
packaging==23.2
pluggy==1.4.0
pyTelegramBotAPI==4.15.3
pytest==8.0.0
requests==2.31.0
urllib3==2.1.0
